fix: cannot test http latency

+++ service/core/v2ray/transparent.go
@@ -139,5 +139,8 @@ func CheckAndSetupTransparentProxy(checkRunning bool) (err error) {
 }
 
 func CheckAndStopTransparentProxy() {
+	if conf.GetEnvironmentConfig().Lite {
+		return
+	}
 	DeleteTransparentProxyRules()
 }
